structure: move feature update review to folder u and create comment to folder c
