+++ src/app/title/[slug]/comments/@components/crud/r.comment.tsx
@@ -1,5 +1,6 @@
 "use client";
-import CComment from "@/app/title/[slug]/comments/@components/crud/c.comment";
+
+import CComment from "@/app/title/[slug]/comments/@components/crud/c/c.comment";
 import RCommentList from "@/app/title/[slug]/comments/@components/crud/r.comment.list";
 import { sendRequest } from "@/utils/api";
