+++ src/app/title/[slug]/reviews/@components/crud/r.review.comments.tsx
@@ -4,13 +4,13 @@ import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Separator } from "@/components/ui/separator";
 import RatingReadOnly from "@/app/title/[slug]/reviews/@components/rating.read.only";
 import CReviewComments from "@/app/title/[slug]/reviews/@components/crud/c.review.comments";
-import UReviewComments from "@/app/title/[slug]/reviews/@components/crud/u.review.comments";
+import UReviewComments from "@/app/title/[slug]/reviews/@components/crud/u/u.review.comments";
 import DReviewComments from "@/app/title/[slug]/reviews/@components/crud/d.review.comments";
 import { sendRequest } from "@/utils/api";
 import Link from "next/link";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
-import BtnUReview from "@/app/title/[slug]/reviews/@components/crud/btn.u.review";
+import BtnUReview from "@/app/title/[slug]/reviews/@components/crud/u/btn.u.review";
 
 dayjs.extend(relativeTime);
